pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-apple-music"
version = "0.1.0"
description = "MCP Server for Apple Music — search, library, playlist management via Claude"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["mcp", "apple-music", "musickit", "claude", "anthropic"]
authors = [
    { name = "Mario Inghilleri", email = "mario@airbagstudio.it" },
    { name = "Claude (Anthropic)", email = "claude@anthropic.com" },
]

dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
]

[project.scripts]
mcp-apple-music       = "mcp_apple_music.server:main"
mcp-apple-music-setup = "mcp_apple_music.setup:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_apple_music"]
